% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{burn_trans_model}
\alias{burn_trans_model}
\title{burn_trans_model}
\usage{
burn_trans_model(
  N_draw,
  trans_parms,
  burn_time = 6000,
  burn_batch_size = 100,
  ncores = 1,
  path = ".",
  save_output = FALSE
)
}
\arguments{
\item{N_draw}{Number of simulated runs.}

\item{trans_parms}{Transmission model parameters.}

\item{burn_time}{Number of months to burn-in model. Default is 6000 (equivalent to 500 years).}

\item{burn_batch_size}{Size of each burn_time batch to run in parallel. Default is 100. Reduce to manage memory workload.}

\item{ncores}{Number of cores to run in parallel. Default is one.}

\item{path}{Path to the location of the Data and Parameters folders. Default is "." and may be used if the function is run inside the package.}

\item{save_output}{Logical. If set to TRUE, then the generated distributions will be saved into the Parameters folder. Default is FALSE.}
}
\value{
Array of estimated model states after burn-in to use as initial values.
}
\description{
Runs the base transmission model for a set burn-in period.
}
