% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_functions.R
\name{cea_parms_age_helper}
\alias{cea_parms_age_helper}
\title{cea_parms_age_helper}
\usage{
cea_parms_age_helper(N_draw, par_nm)
}
\arguments{
\item{N_draw}{Number of parameter draws to characterise distributions.}

\item{par_nm}{Name of the parameter to simulate.}
}
\value{
Named matrix containing age-specific parameter distributions.
}
\description{
Helper function to generate the health economic parameter distributions. Allows for age-specific distributions.
}
