% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{draw_beta}
\alias{draw_beta}
\title{draw_beta}
\usage{
draw_beta(pars, N_draw)
}
\arguments{
\item{pars}{Matrix of hyperparameters with rows for the hyperparameters "a" and "b", and columns for the arms.}

\item{N_draw}{Number of samples to draw.}
}
\value{
Matrix containing N_draw samples drawn (rows) for each arm (columns).
}
\description{
Draws samples from the Beta distribution given a set of hyperparameters.
}
\examples{
pars <- matrix(c(1, 2, 1, 3), nrow = 2, ncol = 2, dimnames = list(c("a", "b"), c("NS", "MV")))
N_draw <- 10000

draws <- draw_beta(pars, N_draw)
colMeans(draws)
}
