% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_functions.R
\name{eval_cea}
\alias{eval_cea}
\title{eval_cea}
\usage{
eval_cea(N_draw, inc_parms, cea_parms, ages, minimal = FALSE)
}
\arguments{
\item{N_draw}{Number of parameter draws to use. Should match the number of simulations used to estimate the incidence parameters.}

\item{inc_parms}{RSV incidence parameter distributions.}

\item{cea_parms}{Health economic parameter distributions.}

\item{ages}{Vector of ages in years used for transmission modelling.}

\item{minimal}{Logical indicator. If set to TRUE, then only minimal output is stored (i.e., no states). Default is FALSE.}
}
\value{
List containing arrays for states (number of individuals in each health economic state), costs and QALYs.
}
\description{
Evaluates the health economic model.
}
