% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_functions.R
\name{eval_icer}
\alias{eval_icer}
\title{eval_icer}
\usage{
eval_icer(costs, qalys, curr, new)
}
\arguments{
\item{costs}{List containing arrays of costs (output from eval_cea() function).}

\item{qalys}{List containing arrays of QALYs (output from eval_cea() function).}

\item{curr}{Name of strategy to use as the comparator (e.g., the currently implemented strategy).}

\item{new}{Name of the strategy to use as the new strategy (e.g., to compare against the currently implemented strategy).}
}
\value{
List containing incremental costs, incremental QALYs and ICERs.
}
\description{
Evaluates the incremental cost-effectiveness ratio (ICER) for the health economic model.
}
