% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_functions.R
\name{eval_INMB}
\alias{eval_INMB}
\title{eval_INMB}
\usage{
eval_INMB(cea, WTP, ref = "NS", years = NULL)
}
\arguments{
\item{cea}{List containing costs and QALYs for each strategy (output from eval_cea() function).}

\item{WTP}{Vector containing willingness-to-pay thresholds.}

\item{ref}{Reference group to use for the increments. Default is "NS".}

\item{years}{Number of years to compute cumulative costs and QALYs. If NULL, then the maximum number of years will be used. Default is NULL.}
}
\value{
Named list containing incremental net monetary benefit for each willingness-to-pay threshold.
}
\description{
Evaluates the incremental net monetary benefit for the health economic model.
}
