% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_functions.R
\name{evpi}
\alias{evpi}
\title{evpi}
\usage{
evpi(INMB)
}
\arguments{
\item{INMB}{Named list containing incremental net monetary benefit for each willingness-to-pay threshold.}
}
\value{
Named vector containing the expected value of perfect information for each willingness-to-pay threshold.
}
\description{
Computes the expected value of perfect information.
}
