% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_functions.R
\name{evppi}
\alias{evppi}
\title{evppi}
\usage{
evppi(INMB, psa, model)
}
\arguments{
\item{INMB}{Named list containing incremental net monetary benefit for each willingness-to-pay threshold.}

\item{psa}{Named matrix of parameter draws for parameters of interest.}

\item{model}{Generalised additive regression model specification (formula). Should be a function of the parameters of interest.}
}
\value{
Named list containing named vector with the expected value of partial perfect information for each willingness-to-pay threshold in addition to a named list of the partial incremental net benefit.
}
\description{
Computes the expected value of partial perfect information using a non-parametric regression approximation method.
}
