% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_functions.R
\name{evsi}
\alias{evsi}
\title{evsi}
\usage{
evsi(
  INMB,
  dat,
  method = "NP",
  model = NULL,
  alpha = NULL,
  hyper_parms = NULL,
  OR = TRUE,
  MCMC = TRUE,
  ncores = 10,
  trans_parms = NULL,
  y0_burn = NULL,
  years = NULL,
  cea_parms = NULL,
  WTP = NULL,
  ref = "NS",
  INMB_partial = NULL
)
}
\arguments{
\item{INMB}{Named list containing incremental net monetary benefit for each willingness-to-pay threshold.}

\item{dat}{Named matrix of simulated data summaries.}

\item{method}{Approximation method. Either NP for non-parametric regression or MM for moment matching. The moment matching method requires the evppi function to be run in advance using the non-parametric regression method to generate INMB_partial. Default is NP.}

\item{model}{Generalised additive regression model specification (formula). Only required for the non-parametric approximation method. Should be a function of variables in dat. Default is NULL.}

\item{alpha}{Power prior contribution. Must be a scalar between zero and 1 (inclusive). Only required for the moment matching method. Default is NULL.}

\item{hyper_parms}{Hyperparameters. Must be a named list ("NS", "MV", "mAbs") with named elements ("a", "b") for "NS" and ("mu", "sigma") for "MV" and "mAbs". Only required for the moment matching method. Default is NULL.}

\item{OR}{Logical. If set to TRUE, the odds ratio model will be used. Otherwise, the relative risk model will be used. Only required for the moment matching method. Default is TRUE.}

\item{MCMC}{Logical. If set to TRUE, posterior will be estimated using Markov Chain Monte-Carlo. Otherwise, a Laplace approximation will be used (not currently available).  Only required for the moment matching method. Default is TRUE.}

\item{ncores}{Number of cores to utilise. Will be used to determine the number of chains to run in parallel. Only required for the moment matching method. Default is ten.}

\item{trans_parms}{Transmission model parameters. Only required for the moment matching method. Default is NULL.}

\item{y0_burn}{Array of initial values for the transmission models. Must be of dimension c(N_draw, 75 (age groups), 5 (SEIR + 1)). Only required for the moment matching method. Default is NULL.}

\item{years}{Number of years to run the transmission models. Only required for the moment matching method. Default is NULL.}

\item{cea_parms}{Health economic parameter distributions. Only required for the moment matching method. Default is NULL.}

\item{WTP}{Vector containing willingness-to-pay thresholds. Only required for the moment matching method. Default is NULL.}

\item{ref}{Reference group to use for the increments when computing the incremental net monetary benefit. Only required for the moment matching method. Default is "NS".}

\item{INMB_partial}{Samples of INMB for the parameters of interest generated from the evppi function using the non-parametric regression approximation method. Only required for the moment matching approximation method. Default is NULL.}
}
\value{
Expected value of sample information
}
\description{
Computes the expected value of sample information using a non-parametric regression or moment matching approximation method.
}
