% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_functions.R
\name{gen_cea_parms}
\alias{gen_cea_parms}
\title{gen_cea_parms}
\usage{
gen_cea_parms(N_draw, seed = NULL, path = ".", save_output = FALSE)
}
\arguments{
\item{N_draw}{Number of parameter draws to characterise distributions.}

\item{seed}{Optional seed for replication. Default is NULL.}

\item{path}{Path to the location of the Data and Parameters folders. Default is "." and may be used if the function is run inside the package.}

\item{save_output}{Logical. If set to TRUE, then the generated distributions will be saved into the Parameters folder. Default is FALSE.}
}
\value{
List containing cost-effectiveness model parameter distributions.
}
\description{
Generates the parameter distributions for the cost-effectiveness model parameters. Requires access to the Data folder containing data for the Australian life tables (life expectancy by age group) and the quality adjusted life year weights (QALYs by age group).
}
