% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{gen_data}
\alias{gen_data}
\title{gen_data}
\usage{
gen_data(N_draw, N, p_alloc, p_d, exact = TRUE)
}
\arguments{
\item{N_draw}{Number of parameter draws to characterise distributions.}

\item{N}{Total trial sample size.}

\item{p_alloc}{Allocation probabilities for each arm. Must be a vector of values between zero and one, and must sum to one.}

\item{p_d}{Named list ("MA", "HP") of matrices containing draws of p_d with columns for p_NS, p_MV and p_mAbs.}

\item{exact}{Logical. If set to TRUE, then the total sample size will be split across the arms according to the proportions defined by p_alloc (equivalent to a blocked design). If set to FALSE then allocations are generated dynamically (i.e., without blocking) and risks imbalance. Default is TRUE.}
}
\value{
Arrays containing simulated data for each trial.
}
\description{
Generates data from a series of trials.
}
