% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{gen_eff_distributions}
\alias{gen_eff_distributions}
\title{gen_eff_distributions}
\usage{
gen_eff_distributions(pars, N_draw = 10000)
}
\arguments{
\item{pars}{Matrix of hyperparameters with rows for the hyperparameters "a" and "b", and columns for the arms.}

\item{N_draw}{Number of samples to draw. Default is 10,000.}
}
\value{
Matrix containing N_draw samples drawn (rows) for each arm (columns).
}
\description{
Generates uncertainty distributions for the strategy effectiveness parameters.
}
\examples{
pars <- matrix(c(2, 1, 1, 2, 1, 5), nrow = 2, ncol = 3, dimnames = list(c("a", "b"), c("NS", "MV", "IM")))

draws <- gen_eff_distributions(pars)
colMeans(draws)
}
