% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{gen_hyper_parms}
\alias{gen_hyper_parms}
\title{gen_hyper_parms}
\usage{
gen_hyper_parms(
  outcome = NULL,
  use_lit = TRUE,
  n = NULL,
  y = NULL,
  seed = NULL,
  N_sim = 1e+07,
  path = ".",
  save_output = FALSE
)
}
\arguments{
\item{outcome}{Optional outcome of interest out of "MA" for medical attendances or "HP" for hospitalisations. Only required if use_lit == TRUE. Default is NULL.}

\item{use_lit}{Logical. If set to TRUE, the data from the literature will be used to inform the hyperparameters. Otherwise, user supplied sample sizes (n) and events (y) are required. Default is TRUE.}

\item{n}{Optional named vector of participant sample sizes. Names must be NS, MV and mAbs. Default is NULL.}

\item{y}{Optional named vector of events. Names must be NS, MV and mAbs. Default is NULL.}

\item{seed}{Optional seed for replication. Default is NULL.}

\item{N_sim}{Number of samples to draw to estimate the relative risk distributions for the moment-matching step. Default is 10,000,000.}

\item{path}{Path to the location of the Parameters folder. Only required if saving the output. Default is "." and may be used if the function is run inside the package.}

\item{save_output}{Logical. If set to TRUE, then the generated distributions will be saved into the Parameters folder. Default is FALSE.}
}
\value{
Named matrix containing hyperparameters.
}
\description{
Generates hyperparameters for the statistical model parameter prior distributions using a summary power prior method.
}
