% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{gen_outcomes}
\alias{gen_outcomes}
\title{gen_outcomes}
\usage{
gen_outcomes(rand, p_d)
}
\arguments{
\item{rand}{Randomisation list generated by gen_randomisation().}

\item{p_d}{Vector of outcome probabilities for each arm. Values must be between zero and one.}
}
\value{
Vector of outcomes of equal length to rand.
}
\description{
Generates Bernoulli outcome data.
}
\examples{
N <- 1000
p_alloc <- c(0.5, 0.5)
rand <- gen_randomisation(N, p_alloc, exact = TRUE)
p_d <- c(0.2, 0.3)

out <- gen_outcomes(rand, p_d)
prop.table(table(rand, out), 1)
}
