% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{gen_randomisation}
\alias{gen_randomisation}
\title{gen_randomisation}
\usage{
gen_randomisation(N, p_alloc, exact)
}
\arguments{
\item{N}{Total sample size.}

\item{p_alloc}{Allocation probabilities for each arm. Must be a vector of values between zero and one, and must sum to one.}

\item{exact}{Logical. If set to TRUE, then the total sample size will be split across the arms according to the proportions defined by p_alloc (equivalent to a blocked design). If set to FALSE then allocations are generated dynamically (i.e., without blocking) and risks imbalance.}
}
\value{
Vector of arm allocations of length N.
}
\description{
Generates a randomisation list.
}
\examples{
N <- 1000
p_alloc <- c(0.5, 0.5)

rand_exact <- gen_randomisation(N, p_alloc, exact = TRUE)
table(rand_exact)

rand_inexact <- gen_randomisation(N, p_alloc, exact = FALSE)
table(rand_inexact)
}
