% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{log_marginal_post}
\alias{log_marginal_post}
\title{log_marginal_post}
\usage{
log_marginal_post(log_r, n, y, hyper_parms, alpha, N_grid)
}
\arguments{
\item{log_r}{Proposal vector for the log relative risks (length two).}

\item{n}{Vector of participant sample sizes.}

\item{y}{Vector of events.}

\item{hyper_parms}{Hyperparameters. Must be a named list ("NS", "MV", "mAbs") with named elements ("a", "b") for "NS" and ("mu", "sigma") for "MV" and "mAbs".}

\item{alpha}{Power prior contribution. Must be a scalar between zero and 1 (inclusive).}

\item{N_grid}{Number of grid points to sample from to evaluate the integral.}
}
\value{
The log marginal joint posterior evaluated at log_r.
}
\description{
Computes the log marginal joint posterior distribution.
}
