% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mean_and_quant_to_pars}
\alias{mean_and_quant_to_pars}
\title{mean_and_quant_to_pars}
\usage{
mean_and_quant_to_pars(dist, mn, quant, pb, N_sim)
}
\arguments{
\item{dist}{Either "beta" or "gamma" to specify distribution.}

\item{mn}{Distribution mean.}

\item{quant}{Distribution quantile.}

\item{pb}{Probability associated with quantile.}

\item{N_sim}{Number of simulations used to validate solution.}
}
\value{
List containing parameter solution (shape and rate parameters if Gamma) with sampled mean and interval to validate solution.
}
\description{
Computes parameters for either a Beta or Gamma distribution given mean and single quantile. Uses a normal approximation and method of moments to generate reasonable initial values then solves numerically.
}
