% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_functions.R
\name{plot_cea}
\alias{plot_cea}
\title{plot_cea}
\usage{
plot_cea(cea_increm, WTP = NULL, sub = NULL)
}
\arguments{
\item{cea_increm}{List containing incremental costs and incremental QALYs for each comparison (output from eval_icer() function).}

\item{WTP}{Vector containing willingness-to-pay thresholds to plot. If NULL, then no thresholds will be included. Default is NULL.}

\item{sub}{Vector containing subset of comparisons to visualise. If NULL, then all comparisons will be included. Default is NULL.}
}
\value{
ggplot object visualising the incremental costs against the incremental QALYs for each comparison.
}
\description{
Plots the distribution of the incremental costs against the incremental QALYs for the health economic model.
}
