% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voi_functions.R
\name{prep_dat_evsi}
\alias{prep_dat_evsi}
\title{prep_dat_evsi}
\usage{
prep_dat_evsi(p_d, N, p_alloc, MM = FALSE, Q = NULL, exact = TRUE)
}
\arguments{
\item{p_d}{Named list ("MA", "HP") of matrices containing draws of p_d with columns for p_NS, p_MV and p_mAbs.}

\item{N}{Total trial sample size.}

\item{p_alloc}{Allocation probabilities for each arm. Must be a vector of values between zero and one, and must sum to one.}

\item{MM}{Logical. If set to TRUE, then data will be simulated in preparation for the moment matching approximation method to be used in the EVSI estimation. Default is FALSE.}

\item{Q}{Number of model reruns to estimate the expected variance of the posterior net benefit. Only required if MM == TRUE. Default is NULL.}

\item{exact}{Logical. If set to TRUE, then the total sample size will be split across the arms according to the proportions defined by p_alloc (equivalent to a blocked design). If set to FALSE then allocations are generated dynamically (i.e., without blocking) and risks imbalance. Default is TRUE.}
}
\value{
Expected value of sample information
}
\description{
Prepares data for input into the evsi function.
}
