% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{quants_to_pars}
\alias{quants_to_pars}
\title{quants_to_pars}
\usage{
quants_to_pars(dist, lower, upper, pb, N_sim)
}
\arguments{
\item{dist}{Either "beta" or "gamma" to specify distribution.}

\item{lower}{Lower bound of interval.}

\item{upper}{Upper bound of interval.}

\item{pb}{Vector containing lower and upper probability bounds.}

\item{N_sim}{Number of simulations used to validate solution.}
}
\value{
List containing parameter solution (shape and rate parameters if Gamma) with sampled mean and interval to validate solution.
}
\description{
Computes parameters for either a Beta or Gamma distribution given an interval. Uses a normal approximation and method of moments to generate reasonable initial values then solves numerically.
}
