% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{run_trans_model}
\alias{run_trans_model}
\title{run_trans_model}
\usage{
run_trans_model(
  N_draw,
  y0,
  mod,
  trans_parms,
  years,
  batch_size = 100,
  ncores = 1
)
}
\arguments{
\item{N_draw}{Number of simulated runs.}

\item{y0}{Array of initial values. Must be of dimension c(N_draw, 75 (age groups), 4 SEIR)).}

\item{mod}{Transmission model to run. Either "base", "vax" or "mab".}

\item{trans_parms}{Transmission model parameters.}

\item{years}{Number of years to run the model.}

\item{batch_size}{Size of each batch for parallel simulations. Default is 100.}

\item{ncores}{Number of cores to run in parallel. Default is one.}
}
\value{
Incidence aggregated by year.
}
\description{
Runs a transmission model and extracts the incidence outputs.
}
