% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{sim_trial}
\alias{sim_trial}
\title{sim_trial}
\usage{
sim_trial(N, p_alloc, p_k, priors, exact = TRUE)
}
\arguments{
\item{N}{Total sample size.}

\item{p_alloc}{Allocation probabilities for each arm. Must be a vector of values between zero and one, and must sum to one.}

\item{p_k}{Vector of outcome probabilities for each arm. Values must be between zero and one.}

\item{priors}{Prior distribution hyperparameters. Must be a matrix with rows for the hyperparameters "a" and "b", and columns for the arms.}

\item{exact}{Logical. If set to TRUE, then the total sample size will be split across the arms according to the proportions defined by p_alloc (equivalent to a blocked design). If set to FALSE then allocations are generated dynamically (i.e., without blocking) and risks imbalance. Default is TRUE.}
}
\value{
Matrix of updated hyperparameters with the same dimension names as the input priors.
}
\description{
Simulates a trial by generating a randomisation list, generating the outcomes and updating the Beta-Binomial posterior distribution.
}
\examples{
N <- 1000
p_alloc <- c(0.5, 0.5)
p_k <- c(0.2, 0.3)
priors <- matrix(rep(1, 4), nrow = 2, ncol = 2, dimnames = list(c("a", "b"), c("NS", "MV")))

sim_trial(N, p_alloc, p_k, priors)
}
