% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{sub_trans_parms}
\alias{sub_trans_parms}
\title{sub_trans_parms}
\usage{
sub_trans_parms(trans_parms, start, end)
}
\arguments{
\item{trans_parms}{Full set of transmission parameter distributions.}

\item{start}{First draw to subset.}

\item{end}{Final draw to subset.}
}
\value{
Subsetted list containing transmission model parameter distributions.
}
\description{
Subsets the parameter distributions for the transmission model parameters. Includes strategy effectiveness parameters.
}
