% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{upt_posterior}
\alias{upt_posterior}
\title{upt_posterior}
\usage{
upt_posterior(n, y, alpha, hyper_parms, OR, N_draw, MCMC = TRUE, nchains = 8)
}
\arguments{
\item{n}{Vector of participant sample sizes.}

\item{y}{Vector of events.}

\item{alpha}{Power prior contribution. Must be a scalar between zero and 1 (inclusive).}

\item{hyper_parms}{Hyperparameters. Must be a named list ("NS", "MV", "mAbs") with named elements ("a", "b") for "NS" and ("mu", "sigma") for "MV" and "mAbs".}

\item{OR}{Logical. If set to TRUE, the odds ratio model will be used. Otherwise, the relative risk model will be used.}

\item{N_draw}{Number of samples to draw.}

\item{MCMC}{Logical. If set to TRUE, posterior will be estimated using Markov Chain Monte-Carlo. Otherwise, a Laplace approximation will be used (not currently available). Default is TRUE.}

\item{nchains}{Number of chains to run in parallel.}
}
\value{
Named matrix of posterior draws.
}
\description{
Updates the posterior distribution using Markov Chain Monte-Carlo.
}
