% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{gen_prior_parms}
\alias{gen_prior_parms}
\title{gen_prior_parms}
\usage{
gen_prior_parms(N_draw, hyper_parms, alpha, seed = NULL)
}
\arguments{
\item{N_draw}{Number of parameter draws to characterise distributions.}

\item{hyper_parms}{Hyperparameters. Must be a named list ("NS", "MV", "mAbs") with named elements ("a", "b") for "NS" and ("mu", "sigma") for "MV" and "mAbs".}

\item{alpha}{Power prior contribution. Must be a scalar between zero and 1 (inclusive).}

\item{seed}{Optional seed for replication. Default is NULL.}
}
\value{
Statistical model prior parameter distributions.
}
\description{
Generates the (prior) parameter distributions for the statistical model parameters.
}
