% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{gen_trans_parms}
\alias{gen_trans_parms}
\title{gen_trans_parms}
\usage{
gen_trans_parms(N_draw, seed = NULL, path = ".", save_output = FALSE)
}
\arguments{
\item{N_draw}{Number of parameter draws to characterise distributions.}

\item{seed}{Optional seed for replication. Default is NULL.}

\item{path}{Path to the location of the Data and Parameters folders. Default is "." and may be used if the function is run inside the package.}

\item{save_output}{Logical. If set to TRUE, then the generated distributions will be saved into the Parameters folder. Default is FALSE.}
}
\value{
List containing transmission model parameter distributions.
}
\description{
Generates the parameter distributions for the transmission model parameters. Excludes strategy effectiveness parameters. Requires access to the Data folder containing data for the Australian population (size by age group) and the estimated mixing matrix (computed via the RSVModels package).
}
