% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{run_trans_models}
\alias{run_trans_models}
\title{run_trans_models}
\usage{
run_trans_models(N_draw, y0, trans_parms, years, ncores = 1)
}
\arguments{
\item{N_draw}{Number of simulated runs.}

\item{y0}{Array of initial values. Must be of dimension c(N_draw, 75 (age groups), 5 (SEIR + 1)).}

\item{trans_parms}{Transmission model parameters.}

\item{years}{Number of years to run the model.}

\item{ncores}{Number of cores to run in parallel. Default is one.}
}
\value{
Incidence and administration aggregated by year.
}
\description{
Runs the transmission models for each strategy and extracts the incidence outputs and administration outputs (where applicable).
}
