% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{upt_posterior}
\alias{upt_posterior}
\title{upt_posterior}
\usage{
upt_posterior(pars, rand, out)
}
\arguments{
\item{pars}{Current set of hyperparameters (usually from the prior distribution). Must be a matrix with rows for the hyperparameters "a" and "b", and columns for the arms.}

\item{rand}{Randomisation list generated by gen_randomisation().}

\item{out}{Vector of outcomes generatedby gen_outcomes().}
}
\value{
Matrix of updated hyperparameters with the same dimension names as the input hyperparameters.
}
\description{
Updates a Beta-Binomial posterior distribution.
}
\examples{
pars <- matrix(rep(1, 4), nrow = 2, ncol = 2, dimnames = list(c("a", "b"), c("NS", "MV")))
N <- 1000
p_alloc <- c(0.5, 0.5)
rand <- gen_randomisation(N, p_alloc, exact = TRUE)
p_k <- c(0.2, 0.3)
out <- gen_outcomes(rand, p_k)

upt_posterior(pars, rand, out)
}
