% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{upt_posterior}
\alias{upt_posterior}
\title{upt_posterior}
\usage{
upt_posterior(
  n,
  y,
  hyper_parms,
  alpha,
  init_vals = c(0, 0),
  N_draw = 10000,
  N_grid = 10000
)
}
\arguments{
\item{n}{Vector of participant sample sizes.}

\item{y}{Vector of events.}

\item{hyper_parms}{Hyperparameters. Must be a named list ("NS", "MV", "mAbs") with named elements ("a", "b") for "NS" and ("mu", "sigma") for "MV" and "mAbs".}

\item{alpha}{Power prior contribution. Must be a scalar between zero and 1 (inclusive).}

\item{init_vals}{Vector containing initial values for optimisation. Default is c(0, 0).}

\item{N_draw}{Number of samples to draw. Default is 10,000.}

\item{N_grid}{Number of grid points to sample from to evaluate the integral. Default is 10,000.}
}
\value{
Named matrix of posterior draws for the relative risk parameters.
}
\description{
Updates the posterior distribution using a Laplace approximation.
}
